server/requirements.txt
flask
flask-cors
tldextract
pillow